feat(app): move Guard folder to lib/components
use AppErrorBoundary
