+++ src/client/Router.tsx
@@ -3,8 +3,8 @@ import { useHashLocation } from "./useHashLocation";
 
 import { RouteAccess, routes } from "./config/routes";
 
-import { PrivatePage } from "./ui/molecules/Guard/PrivatePage";
-import { AnonymousPage } from "./ui/molecules/Guard/AnonymousPage";
+import { PrivatePage } from "./lib/components/Guard/PrivatePage";
+import { AnonymousPage } from "./lib/components/Guard/AnonymousPage";
 import { MainLayout } from "./layouts/MainLayout/MainLayout";
 import { DeepLinkHandler } from "./DeepLinkHandler";
 import { NotFoundPage } from "./pages/NotFound/NotFoundPage";
