+++ src/client/pages/MySaves/MySavesPage.tsx
@@ -2,8 +2,8 @@ import { useState } from "react";
 
 import { H1 } from "@/client/ui/atoms/Typography";
 import { Container } from "@/client/ui/atoms/Container/Container";
-import { FolderExplorer } from "@/client/ui/organisms/FolderExplorer/FolderExplorer";
-import { MySavesWidget } from "@/client/ui/organisms/MySavesWidget/MySavesWidget";
+import { FolderExplorer } from "@/client/pages/MySaves/FolderExplorer/FolderExplorer";
+import { MySavesWidget } from "@/client/pages/MySaves/MySavesWidget/MySavesWidget";
 
 export const MySavesPage = () => {
   const [onSaveUpload, setOnSaveUpload] = useState(() => ({
